addressing (some) review comments
